Adding documentation to macro for statistics + moving to statistics crate
